+++ app/src/main/java/com/seom/seommain/data/repository/MailRepository.kt
@@ -4,23 +4,9 @@ import com.seom.seommain.data.entity.MailEntity
 import java.lang.Exception
 import java.util.*
 
-class MailRepository {
-
-    fun getAllMail(): Result<List<MailEntity>> =
-        try {
-            val mockData = (0..100).map {
-                MailEntity(
-                    id = it,
-                    sender = if (it % 2 == 0) "${it}th sender" else "sender ${it}th",
-                    title = "title $it",
-                    content = "content $it",
-                    date = Date(),
-                    type = it % 3
-                )
-            }
-
-            Result.success(mockData.toList())
-        } catch (exception: Exception) {
-            Result.failure(exception)
-        }
+interface MailRepository {
+    /**
+     * 모든 메일 정보 요청
+     */
+    suspend fun getAllMail(): Result<List<MailEntity>>
 }
