+++ app/src/main/java/com/seom/seommain/di/RepositoryModule.kt
@@ -1,0 +1,17 @@
+package com.seom.seommain.di
+
+import com.seom.seommain.data.repository.MailRepository
+import com.seom.seommain.data.repository.impl.MailRepositoryImpl
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class RepositoryModule {
+    @Binds
+    abstract fun bindMailRepository(
+        mailRepositoryImpl: MailRepositoryImpl
+    ): MailRepository
+}
