+++ app/src/main/java/com/seom/seommain/ui/home/HomeViewModel.kt
@@ -4,10 +4,13 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.seom.seommain.ui.model.BaseType
 import com.seom.seommain.ui.model.mail.MailType
+import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import javax.inject.Inject
 
-class HomeViewModel : ViewModel() {
+@HiltViewModel
+class HomeViewModel @Inject constructor() : ViewModel() {
 
     // mail 타입
     private val _drawerSelectedType = MutableStateFlow<BaseType>(MailType.PRIMARY)
